[CF] round 2 changes
